+++ CHANGELOG.md
@@ -38,6 +38,7 @@ noticeable to end-users since the last release. For developers, this project fol
   ([#30](https://github.com/loichyan/nerdfix/pull/30)).
 - Support comments in JSON files.
 - Add a new `nerdfix query` subcommand, useful for querying icon infos from the database.
+- Add a new `codepoint:from/to` substitution type.
 
 ## [0.4.1] - 2024-07-14